Final level 0 background and Kallum stands still when getting to an open door
I've added shadows to the letters in level 0 so as to make the platforms more visible, but I don't get why the background still looks so pixelated. I've also modified some scripts a bit in order to make Kallum stand still when he gets to the end of the level (before, it was possible for the player to jump to the bottom of the level or go away during the few seconds where the game is about to transition to the next level).